+++ src/frformat/geo/canton.py
@@ -1,7 +1,7 @@
 from typing import Dict, FrozenSet
 
 from frformat import geo_enum_format
-from frformat.geo.canton_set import (
+from frformat.geo.canton_frozenset import (
     CANTON_FROZEN_SET_COG_2023,
     CANTON_FROZEN_SET_COG_2024,
 )
