+++ src/frformat/geo/canton_frozenset.py
