+++ src/frformat/geo/code_commune_insee_frozenset.py
