+++ src/frformat/geo/code_fantoir.py
@@ -2,7 +2,7 @@ import string
 
 from frformat import CustomStrFormat, Metadata
 from frformat.formatter import Formatter
-from frformat.geo.code_fantoir_set import PARTIAL_CODE_FANTOIR_FROZEN_SET
+from frformat.geo.code_fantoir_frozenset import PARTIAL_CODE_FANTOIR_FROZEN_SET
 
 name = "Code fantoir"
 description = "Vérifie les codes fantoirs valides"
