+++ src/frformat/geo/code_fantoir_frozenset.py
