+++ src/frformat/geo/code_postal_frozenset.py
