+++ src/frformat/geo/commune_frozenset.py
