+++ src/frformat/geo/departement_frozenset.py
