+++ src/frformat/geo/pays.py
@@ -1,7 +1,7 @@
 from typing import Dict, FrozenSet
 
 from frformat import geo_enum_format
-from frformat.geo.pays_set import PAYS_FROZEN_SET_COG_2024
+from frformat.geo.pays_frozenset import PAYS_FROZEN_SET_COG_2024
 from frformat.geo_enum_format import Millesime
 
 name = "Pays et territoires étrangers"
