+++ src/frformat/geo/region_frozenset.py
